requirements.txt
SimpleITK
pytorch_lightning
monai
wandb
torch-summary
einops